Initial version
This simple python script, will take the service name as the argument number 1 and check using ps - C command, if that process is running. If it is not, then it will try to restart the process in one attempt. If the restart fails, it will write an error to the log.

This module uses the subprocess module to make the call to unix system.

You can always cron this to make sure that the server is always running with the process.